third_party/blink/web_tests/http/tests/misc/extract-http-content-language-against-equiv-expected.txt

Test for bug 97929: Extract HTTP Content-Language header.

==> Value extracted from meta "http-equiv" and overriding HTTP "Content-Language"...
PASS window.getComputedStyle(document.getElementById("x")).webkitLocale is '"en-GB"'
==> Value set by div "lang" tag...
PASS window.getComputedStyle(document.getElementById("y")).webkitLocale is '"ar"'
==> All done...
PASS successfullyParsed is true

TEST COMPLETE
